# openapi.yaml
openapi: 3.0.3
info:
  title: Suminregio API
  description: |
    API para dashboard ejecutivo de Suminregio Parker (ventas, CXC, inventario,
    contabilidad, consumos, análisis avanzado). Consume DuckDB snapshots.
  version: '9.1'
  contact:
    name: Guillermo Ragnar
servers:
  - url: https://suminregioparker-todoslosnegocios.onrender.com
    description: Producción (Render)
  - url: http://localhost:7000
    description: Desarrollo local
...
paths:
...
  /api/reports/preview:
    get:
      tags: [Reports]
      summary: HTML del reporte diario (sin enviar email)
      parameters:
        - in: query
          name: db
          schema: { type: string }
  /api/reports/send:
    post:
      tags: [Reports]
      summary: Envío manual de reporte por email
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                to: { type: string, format: email }
                subject: { type: string }
                db: { type: string }
...